+++ tests/link/link.spec.js
@@ -1,0 +1,39 @@
+import AxeBuilder from '@axe-core/playwright';
+import percySnapshot from '@percy/playwright';
+import { expect } from '@playwright/test';
+import { test } from '../base-fixture';
+
+test.describe('Link tests', () => {
+  const url = '/components/link/example-index.html';
+
+  test.beforeEach(async ({ page }) => {
+    await page.goto(url);
+  });
+
+  test.describe('general page checks', () => {
+    test('should have a title', async ({ page }) => {
+      await expect(page).toHaveTitle('IDS Enterprise');
+    });
+  });
+
+  test.describe('accessibility tests', () => {
+    test('should pass an Axe scan', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+      const accessibilityScanResults = await new AxeBuilder({ page })
+        .disableRules(['meta-viewport', 'landmark-main-is-top-level', 'landmark-no-duplicate-main', 'nested-interactive'])
+        .exclude('[disabled]')
+        .analyze();
+      expect(accessibilityScanResults.violations).toEqual([]);
+    });
+  });
+
+  test.describe('snapshot tests', () => {
+    test('should match the visual snapshot in percy', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+      await percySnapshot(page, 'link-light');
+    });
+  });
+
+  test.describe('functionality tests', () => {
+  });
+});